- Optimized database queries in training sessions retrival that caused n+1.

+++ src/main/java/com/gymtracker/training_session/TrainingSessionResponseDto.java
@@ -7,6 +7,6 @@ import java.time.LocalDate;
 public record TrainingSessionResponseDto(Long id, String routineName,
                                          @JsonFormat(pattern = "dd-MM-yyyy")
                                          LocalDate trainingDate,
-                                         Integer totalWeight) {
+                                         Long totalWeight) {
 
 }
