+++ src/main/java/com/gymtracker/training_session/progress_tracker/ProgressTrackerService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface ProgressTrackerService {
     List<TrainingLogResponseDto> sortByMaxWeightAndRepsForExercise(Long exerciseId);
 
-    Integer calculateTotalWeightForSession(Long trainingSessionId);
+    Long calculateTotalWeightForSession(Long trainingSessionId);
 
-    Integer calculateExerciseWeightForSession(Long exerciseId,
+    Long calculateExerciseWeightForSession(Long exerciseId,
                                               Long trainingSessionId);
 }
